Update customer log
version 1
