+++ MerlionBank-ejb/src/java/CustomerRelationshipEntity/Session/OperationalCRMSessionBean.java
@@ -1,0 +1,19 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package CustomerRelationshipEntity.Session;
+
+import javax.ejb.Stateless;
+
+/**
+ *
+ * @author a0113893
+ */
+@Stateless
+public class OperationalCRMSessionBean implements OperationalCRMSessionBeanLocal {
+
+    // Add business logic below. (Right-click in editor and choose
+    // "Insert Code > Add Business Method")
+}
