+++ MerlionBank-ejb/src/java/CustomerRelationshipEntity/Session/OperationalCRMSessionBeanLocal.java
@@ -1,0 +1,17 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package CustomerRelationshipEntity.Session;
+
+import javax.ejb.Local;
+
+/**
+ *
+ * @author a0113893
+ */
+@Local
+public interface OperationalCRMSessionBeanLocal {
+    
+}
